add indicator for loading preview
move render API call to preview so preview can spin the refresh
icon at the right time.

Also this diff add an animated loading page
