+++ editor/src/customize.css
@@ -231,3 +231,42 @@ iframe.preview {
 .toast {
     border-radius: 10px;
 }
+
+
+.Loading {
+    height: 100%;
+    width: 100%;
+    background-image: url('./icons/logo.svg');
+    background-repeat:no-repeat;
+    background-position: center center;
+    background-size: 100px 100px;
+    animation-name: fade;
+    animation-duration: 5s;
+    animation-iteration-count: infinite;
+    transform-origin: 50% 50%;
+}
+@keyframes fade {
+    0% {
+        opacity: 0.05;
+    }
+    50% {
+        opacity: 0.2;
+    }
+    100% {
+        opacity: 0.05;
+    }
+}
+.spin{
+    animation-name: spin;
+    animation-duration: 3s;
+    animation-iteration-count: infinite;
+    transform-origin: 50% 50%;
+}
+@keyframes spin {
+    0% {
+        transform: rotate(0deg);
+    }
+    100% {
+        transform: rotate(360deg);
+    }
+}
